refactor: Add useParkingLot hook to fetch parking lot data
This commit adds a new hook called useParkingLot to fetch parking lot data in the Home.LotDetailModal component. The hook uses the trpc.parking.lot.get.single.useQuery function to fetch the data based on the selectedLotId. The fetched data is then used to display the name of the parking lot in a Button component. If the data is still being fetched, a loading indicator is shown instead.